implement Ord for Type
